Refactor dormitory pagination handling: introduce DormitoryPaginationDto for improved data structure; update DormitoryService and DormitoryController to utilize new DTO; add FileRepository for image retrieval.

+++ src/main/java/com/tam/StudentManagement/Model/File.java
@@ -35,5 +35,5 @@ public class File extends BaseModel {
     private String fileableType;
 
     @Column(name = "fileable_id")
-    private Long fileableId;
+    private Integer fileableId;
 }
